# add (before|around|after)_render hooks