--- a/lib/action_view/component.rb
+++ b/lib/action_view/component.rb
@@ -13,6 +13,7 @@ module ActionView
     autoload :Conversion
     autoload :Preview
     autoload :Previewable
+    autoload :Callbacks
     autoload :TestHelpers
     autoload :TestCase
     autoload :RenderMonkeyPatch