--- a/test/app/components/callbacks_component.html.erb
+++ b/test/app/components/callbacks_component.html.erb
@@ -1,0 +1,1 @@
+Rendered