--- a/test/app/components/conditional_render_component.rb
+++ b/test/app/components/conditional_render_component.rb
@@ -1,6 +1,7 @@
 # frozen_string_literal: true
 
 class ConditionalRenderComponent < ActionView::Component::Base
+<<<<<<< HEAD
   validates :should_render, inclusion: { in: [true, false] }
 
   def initialize(should_render:)
@@ -15,5 +16,16 @@ class ConditionalRenderComponent < ActionView::Component::Base
     end
 
     should_render
+=======
+  before_render do
+    self.rendered_output = "" unless should_render
+  end
+
+  attr_reader :should_render
+  private :should_render
+
+  def initialize(should_render:)
+    @should_render = should_render
+>>>>>>> add (before|around|after)_render hooks
   end
 end