--- a/test/app/components/overrided_render_component.html.erb
+++ b/test/app/components/overrided_render_component.html.erb
@@ -1,0 +1,1 @@
+Cache <%= version %>