--- a/test/app/components/render_check_component.rb
+++ b/test/app/components/render_check_component.rb
@@ -1,6 +1,10 @@
 # frozen_string_literal: true
 
 class RenderCheckComponent < ActionView::Component::Base
+  before_render do
+    self.rendered_output = "" unless render?
+  end
+
   def initialize(*); end
 
   def render?