--- a/test/app/views/integration_examples/overrided_render.html.erb
+++ b/test/app/views/integration_examples/overrided_render.html.erb
@@ -1,0 +1,1 @@
+<%= render(OverridedRenderComponent, version: params[:version]) %>