--- a/test/config/routes.rb
+++ b/test/config/routes.rb
@@ -10,4 +10,5 @@ Dummy::Application.routes.draw do
   get :variants, to: "integration_examples#variants"
   get :cached, to: "integration_examples#cached"
   get :render_check, to: "integration_examples#render_check"
+  get :overrided_render, to: "integration_examples#overrided_render"
 end