@jcoyne 2020-01-22
Can you add an explanation of why you are making this change?  A link to #187 is not enough for me to get the context.