@xronos-i-am 2020-01-23
@jcoyne `before_render` seems to be more "rails" way of preventing render. `around_render` can be used for instrumentation for example