@jonspalmer 2020-01-23
I like this approach a great deal. It feels more natural to support halting the rendering this way and allows for instrumentation and a bunch of other extensions. The gap wrt #187 is how to support caching but I think we can solve that. The trick is to stop relying on the return value of the callbacks. How about we add an attr_accessor like `rendered_output` to `Base`. Change the render_template to set it:

```ruby
      def render_template(&block)
        old_current_template = @current_template
        @current_template = self

        @content = view_context.capture(self, &block) if block_given?

        validate!

        rendered_output = send(self.class.call_method_name(@variant))
      ensure
        @current_template = old_current_template
      end
```

and then in the render_in method instead or relying on the return value of run_callbacks we do this

```ruby
run_callbacks :render do
  render_template(&block)
end

rendered_output
```

That way any before, after or around filter could manipulate the rendered_output and everything works.

For example the simple caching version would be:

```
around_render :render_cached
def render_cached
  rendered_output = Rails.cache.fetch(cache_key) do
    yield
  end
end
``` 

I suspect we could also 'automate' filter halting by checking for a non nil value of rendered_output (similar to how Rails Controllers halt the chain if you call `render` or `redirect_to`).