@jcoyne 2020-01-23
@xronos-i-am I'm not criticizing the substance of your PR, I think this is a big improvement, I'm just asking for a better commit message that explains why this work was worth undertaking.  I think that kind of thing can really help future contributors as they search through git history.