@xronos-i-am 2020-01-24
@jonspalmer tried your approach

> I suspect we could also 'automate' filter halting by checking for a non nil value of rendered_output (similar to how Rails Controllers halt the chain if you call `render` or `redirect_to`).

```ruby
before_render do
  self.rendered_output = '' unless render?
end
```

Don't sure it is much better than `throw` or `true/false` resolving

> That way any before, after or around filter could manipulate the rendered_output and everything works.

```ruby
def render_cached
  self.rendered_output = Rails.cache.fetch(cache_key) do
    yield
  end
end
```

It may be confusing. We have to assign `@rendered_output` to return of `fetch` command. It's easy to write the following and hard to find why it doesn't work:

```ruby
def render_cached
  Rails.cache.fetch(cache_key) do
    self.rendered_output = yield
  end
end
```