@xronos-i-am 2020-01-24
#187 should be closed if we use approach with assigning `@rendered_output`