@jonspalmer 2020-01-24
Taking a step back a little bit on the caching question. Why is it important to do the caching in the ruby class vs in the template where we have access to all the usual Rails cache helpers?