@xronos-i-am 2020-01-27
@jonspalmer i want to have minimum overhead when component is cached. There is no need to run additional ruby code which renders a template. Caching whole the component seems to be the same anti-pattern as conditional `render?` in template, which was moved early. I'm pretty sure templates should by as clean as possible, caching logic should be in ActionView::Component class