@rainerborene 2020-01-31
I agree with @jonspalmer. There is something else to consider: ActiveSupport callbacks are slow as hell. By the way, that's **one** of the reasons why `sequel` gem is much more faster than `activerecord`. On my `ApplicationComponent` file I usually redefine the `validate!` to an empty method, as native Ruby keywords works just fine for checking arguments. 

It's worth writing a benchmark code, like rendering a thousand list items, to measure the performance of callbacks before proceeding with this PR (if merged).