@jonspalmer 2020-01-31
@rainerborene do we have a sense of why validate and callbacks are so slow? The code they run to check things is relatively straightforward so what's the overhead they introduce?