@rainerborene 2020-01-31
I didn't have time to go through the AS callback internals. See the results of the benchmark:

```
Warming up --------------------------------------
     with validation     9.901k i/100ms
  without validation   126.863k i/100ms
Calculating -------------------------------------
     with validation    106.268k (± 1.6%) i/s -    534.654k in   5.032420s
  without validation      1.640M (± 3.4%) i/s -      8.246M in   5.035358s

Comparison:
  without validation:  1639630.3 i/s
     with validation:   106268.1 i/s - 15.43x  slower
```

Here is the code of the benchmark itself.

```ruby
require "active_support/all"
require "action_view"
require "action_view/component"
require "benchmark/ips"

class ActionView::Component::Base
  def self.inherited(_child)
    # no routes available here
  end
end

class Row < ActionView::Component::Base
  validates :name, presence: true

  attr_reader :name

  def initialize(name:)
    @name = name
  end
end

Benchmark.ips do |x|
  x.report("with validation") do
    row = Row.new(name: "blah")
    row.validate!
  end

  x.report("without validation") do
    Row.new(name: "blah")
  end

  x.compare!
end
```