@jonspalmer 2020-01-31
Got it. That benchmark is apples to oranges though. My question would be can you do the validation checks manually for equivalent effect but faster than the Rails validations?