@rainerborene 2020-01-31
@jonspalmer There are dozens of alternative validation libraries out there. Need to see which one performs better than `activemodel` and `activesupport` together. But I think our premises are completely wrong here. 

If I need to validate input data from the component perspective which is the rendering layer, then  something is completely wrong with my design. And what if someone wants to use `dry-validation` or something else? You'll need to monkey patch or override a bunch of methods? I vote to wipe out validation from `actionview-component`.

Using validations just to check the presence (calling `blank?`) doesn't worth this unreasonable performance cost for me. If someone needs validation, then just include ActiveModel modules into the base component class. 

I don't think it's necessary to enforce a validation library just to use components at all.