@xronos-i-am 2020-02-03
> I don't think it's necessary to enforce a validation library to use components at all.

totally agree

This PR doesn't have sense for me anymore. Caching of whole component is not possible with merged #196.